fix: allow plain conv layer when no context modulation is used
Previously, the constructor raised an error if both `use_scale` and `use_bias` were False,
even when no context vector was provided. This commit fixes that logic by only requiring
`use_scale` or `use_bias` to be True if `context_dim` is set, allowing use of the layer
as a standard convolution when context modulation is not needed.